Add filesystem abstraction to REST API
The REST API has been updated to allow reading directly from fsspec-supported filesystems. The
new workflow is to first generate metadata on the server via a GET to the '/document' endpoint.
If the metadata references a document on a fsspec-supported path, then we go straight to the
'chats' endpoint. Otherwise, the workflow is as before and we must upload the document to the
server via a POST to the '/document' endpoint.