Fix rendering of chat info button
Currently a new chat info button is created whenever the UI is updated. This
creates a problem as when the UI is re-rendered it is still referencing the
original button (which has not actually been updated). To fix this issue a
button should only be created once and then updated in place prior to each
re-render.